move outboundtrafficpolicy to policy
