+++ pkg/test/framework/features/allowlist.txt
@@ -1509,8 +1509,8 @@ telemetry,TestDashboard/istio-performance-dashboard.json
 telemetry,TestDashboard/istio-service-dashboard.json
 telemetry,TestDashboard/istio-workload-dashboard.json
 telemetry,TestDashboard/pilot-dashboard.json
-telemetry_outboundtrafficpolicy,TestOutboundTrafficPolicy_AllowAny
-telemetry_outboundtrafficpolicy,TestOutboundTrafficPolicy_RegistryOnly
+telemetry_policy,TestOutboundTrafficPolicy_AllowAny
+telemetry_policy,TestOutboundTrafficPolicy_RegistryOnly
 telemetry_requestclassification,TestRequestClassification
 telemetry_stats_prometheus_http_nullvm,TestStatsFilter
 telemetry_stats_prometheus_http_wasm,TestWasmStatsFilter
