+++ tests/integration/telemetry/policy/traffic_allow_any_test.go
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package outboundtrafficpolicy
+package policy
 
 import (
 	"net/http"
